+++ src/assets/icons/edge-extension-provider-icon/edge-extension-provider-icon.tsx
@@ -1,11 +1,11 @@
 import { Component, h, Prop } from '@stencil/core';
 
 @Component({
-  tag: 'mvx-extension-provider-icon',
-  styleUrl: 'extension-provider-icon.scss',
+  tag: 'mvx-edge-extension-provider-icon',
+  styleUrl: 'edge-extension-provider-icon.scss',
   shadow: true,
 })
-export class ExtensionProviderIcon {
+export class EdgeExtensionProviderIcon {
   @Prop() class?: string;
   @Prop() width?: number;
   @Prop() height?: number;
@@ -17,10 +17,10 @@ export class ExtensionProviderIcon {
         viewBox="0 0 46 44"
         width={this.width ?? '46'}
         height={this.height ?? '44'}
-        class={{ 'extension-provider-icon': true, [this.class]: Boolean(this.class) }}
+        class={{ 'edge-extension-provider-icon': true, [this.class]: Boolean(this.class) }}
       >
         <mask
-          id="mask0_7107_23656"
+          id="mask0_10402_51337"
           style={{ maskType: 'alpha' }}
           maskUnits="userSpaceOnUse"
           x="0"
@@ -35,7 +35,7 @@ export class ExtensionProviderIcon {
             fill="white"
           />
         </mask>
-        <g mask="url(#mask0_7107_23656)">
+        <g mask="url(#mask0_10402_51337)">
           <path
             fill-rule="evenodd"
             clip-rule="evenodd"
@@ -49,8 +49,8 @@ export class ExtensionProviderIcon {
             width="39.6602"
             height="39.6602"
             rx="8.76214"
-            fill="url(#paint0_linear_7107_23656)"
-            stroke="url(#paint1_linear_7107_23656)"
+            fill="url(#paint0_linear_10402_51337)"
+            stroke="url(#paint1_linear_10402_51337)"
             stroke-width="0.339806"
           />
           <rect
@@ -59,45 +59,51 @@ export class ExtensionProviderIcon {
             width="39.6602"
             height="39.6602"
             rx="8.76214"
-            fill="url(#paint2_linear_7107_23656)"
-            stroke="url(#paint3_linear_7107_23656)"
+            fill="url(#paint2_linear_10402_51337)"
+            stroke="url(#paint3_linear_10402_51337)"
             stroke-width="0.339806"
           />
-          <g filter="url(#filter0_dddd_7107_23656)">
+          <g filter="url(#filter0_dddd_10402_51337)">
             <path
               d="M13.1387 26.2337L18.8354 23.1782L17.8776 21.3301L12.6614 23.4483C12.527 23.5029 12.3781 23.5029 12.2438 23.4483L7.02607 21.3301L6.06836 23.1782L11.765 26.2337L6.06836 29.2878L7.02607 31.1359L12.2423 29.0177C12.3767 28.9631 12.5256 28.9631 12.6599 29.0177L17.8762 31.1359L18.8339 29.2878L13.1373 26.2323L13.1387 26.2337Z"
               fill="#000F0E"
             />
           </g>
         </g>
-        <g clip-path="url(#clip0_7107_23656)">
+        <g clip-path="url(#clip0_10402_51337)">
           <path
-            d="M38 11.998C40.2091 11.998 42 10.2072 42 7.99805C42 5.78891 40.2091 3.99805 38 3.99805C35.7909 3.99805 34 5.78891 34 7.99805C34 10.2072 35.7909 11.998 38 11.998Z"
-            fill="white"
+            d="M44.441 11.9063C44.2265 12.0178 44.0052 12.1159 43.7785 12.2C43.0607 12.4672 42.3007 12.6027 41.5348 12.6C38.5785 12.6 36.0035 10.5688 36.0035 7.95625C36.0074 7.60643 36.1043 7.26394 36.2845 6.96405C36.4647 6.66415 36.7215 6.41768 37.0285 6.25C34.3535 6.3625 33.666 9.15 33.666 10.7813C33.666 15.4063 37.9223 15.8688 38.841 15.8688C39.3348 15.8688 40.0785 15.725 40.5285 15.5813L40.6098 15.5563C42.3319 14.9611 43.8 13.7972 44.7723 12.2563C44.8005 12.2094 44.8127 12.1546 44.8068 12.1003C44.8009 12.0459 44.7774 11.9949 44.7398 11.9552C44.7022 11.9155 44.6526 11.8892 44.5987 11.8804C44.5447 11.8715 44.4893 11.8806 44.441 11.9063Z"
+            fill="url(#paint4_linear_10402_51337)"
           />
           <path
-            d="M37.9998 4.00001H44.9269C44.225 2.78381 43.2152 1.77385 41.9991 1.07168C40.783 0.369518 39.4035 -9.79568e-05 37.9993 1.94733e-08C36.5951 9.79957e-05 35.2156 0.369906 33.9996 1.07224C32.7837 1.77457 31.774 2.78468 31.0723 4.00098L34.5358 10L34.5389 9.99921C34.1866 9.39164 34.0006 8.70191 33.9998 7.99957C33.9991 7.29722 34.1834 6.60708 34.5344 5.99872C34.8854 5.39035 35.3905 4.88526 35.9989 4.53436C36.6073 4.18346 37.2975 3.99915 37.9998 4.00001Z"
-            fill="url(#paint4_linear_7107_23656)"
+            opacity="0.35"
+            d="M44.441 11.9063C44.2265 12.0178 44.0052 12.1159 43.7785 12.2C43.0607 12.4672 42.3007 12.6027 41.5348 12.6C38.5785 12.6 36.0035 10.5688 36.0035 7.95625C36.0074 7.60643 36.1043 7.26394 36.2845 6.96405C36.4647 6.66415 36.7215 6.41768 37.0285 6.25C34.3535 6.3625 33.666 9.15 33.666 10.7813C33.666 15.4063 37.9223 15.8688 38.841 15.8688C39.3348 15.8688 40.0785 15.725 40.5285 15.5813L40.6098 15.5563C42.3319 14.9611 43.8 13.7972 44.7723 12.2563C44.8005 12.2094 44.8127 12.1546 44.8068 12.1003C44.8009 12.0459 44.7774 11.9949 44.7398 11.9552C44.7022 11.9155 44.6526 11.8892 44.5987 11.8804C44.5447 11.8715 44.4893 11.8806 44.441 11.9063Z"
+            fill="url(#paint5_radial_10402_51337)"
           />
           <path
-            d="M38.0007 11.1668C39.7496 11.1668 41.1673 9.74906 41.1673 8.00016C41.1673 6.25126 39.7496 4.8335 38.0007 4.8335C36.2517 4.8335 34.834 6.25126 34.834 8.00016C34.834 9.74906 36.2517 11.1668 38.0007 11.1668Z"
-            fill="#1A73E8"
+            d="M36.6042 15.0814C36.0479 14.7358 35.5657 14.2833 35.1855 13.7501C34.7513 13.1558 34.4525 12.4738 34.3099 11.7518C34.1673 11.0297 34.1845 10.2853 34.3602 9.57058C34.5359 8.8559 34.8658 8.18833 35.3269 7.61471C35.788 7.04109 36.369 6.57532 37.0292 6.2501C37.2292 6.15635 37.5605 5.99385 38.0042 6.0001C38.3165 6.00251 38.624 6.07711 38.9026 6.21806C39.1813 6.35901 39.4235 6.56249 39.6105 6.8126C39.862 7.15078 39.9998 7.55994 40.0042 7.98135C40.0042 7.96885 41.5355 3.00635 35.0042 3.00635C32.2605 3.00635 30.0042 5.60635 30.0042 7.89385C29.9933 9.10219 30.2517 10.2978 30.7605 11.3938C31.5938 13.1682 33.0496 14.5749 34.8515 15.3468C36.6534 16.1188 38.6761 16.2022 40.5355 15.5814C39.8841 15.7871 39.1957 15.8489 38.5181 15.7626C37.8404 15.6763 37.1895 15.4438 36.6105 15.0814L36.6042 15.0814Z"
+            fill="url(#paint6_linear_10402_51337)"
           />
           <path
-            d="M41.463 10.0009L37.9995 15.9999C39.4037 16.0001 40.7832 15.6306 41.9994 14.9285C43.2155 14.2265 44.2253 13.2166 44.9273 12.0004C45.6294 10.7843 45.9988 9.40474 45.9986 8.00051C45.9983 6.59629 45.6284 5.21687 44.9259 4.00098H37.9989L37.998 4.00408C38.7004 4.00272 39.3907 4.18655 39.9993 4.53704C40.608 4.88752 41.1135 5.39227 41.4649 6.0004C41.8162 6.60854 42.0011 7.29856 42.0008 8.0009C42.0004 8.70325 41.815 9.3931 41.463 10.0009Z"
-            fill="url(#paint5_linear_7107_23656)"
+            opacity="0.41"
+            d="M36.6042 15.0814C36.0479 14.7358 35.5657 14.2833 35.1855 13.7501C34.7513 13.1558 34.4525 12.4738 34.3099 11.7518C34.1673 11.0297 34.1845 10.2853 34.3602 9.57058C34.5359 8.8559 34.8658 8.18833 35.3269 7.61471C35.788 7.04109 36.369 6.57532 37.0292 6.2501C37.2292 6.15635 37.5605 5.99385 38.0042 6.0001C38.3165 6.00251 38.624 6.07711 38.9026 6.21806C39.1813 6.35901 39.4235 6.56249 39.6105 6.8126C39.862 7.15078 39.9998 7.55994 40.0042 7.98135C40.0042 7.96885 41.5355 3.00635 35.0042 3.00635C32.2605 3.00635 30.0042 5.60635 30.0042 7.89385C29.9933 9.10219 30.2517 10.2978 30.7605 11.3938C31.5938 13.1682 33.0496 14.5749 34.8515 15.3468C36.6534 16.1188 38.6761 16.2022 40.5355 15.5814C39.8841 15.7871 39.1957 15.8489 38.5181 15.7626C37.8404 15.6763 37.1895 15.4438 36.6105 15.0814L36.6042 15.0814Z"
+            fill="url(#paint7_radial_10402_51337)"
           />
           <path
-            d="M34.5356 10.001L31.0721 4.00195C30.3698 5.21795 30.0001 6.59742 30 8.00166C29.9999 9.40589 30.3696 10.7854 31.0718 12.0014C31.774 13.2175 32.784 14.2272 34.0002 14.9291C35.2165 15.631 36.596 16.0004 38.0003 16L41.4638 10.0009L41.4616 9.99865C41.1116 10.6076 40.6072 11.1135 39.9994 11.4653C39.3915 11.8172 38.7016 12.0026 37.9993 12.0028C37.2969 12.003 36.6069 11.8181 35.9988 11.4667C35.3908 11.1152 34.8861 10.6097 34.5356 10.001Z"
-            fill="url(#paint6_linear_7107_23656)"
+            d="M39.525 9.3C39.4688 9.3625 39.3125 9.45625 39.3125 9.65C39.3125 9.8125 39.4187 9.975 39.6125 10.1062C40.5062 10.7312 42.2 10.6437 42.2062 10.6437C42.8725 10.6433 43.5265 10.4642 44.1 10.125C44.6774 9.78729 45.1565 9.30452 45.4897 8.72457C45.823 8.14463 45.9989 7.48764 46 6.81875C46.0187 5.41875 45.5 4.4875 45.2937 4.075C43.9625 1.4875 41.1062 -1.4243e-07 38 -9.16404e-08C35.8976 -0.000207945 33.8797 0.827186 32.3826 2.30323C30.8855 3.77928 30.0296 5.78532 30 7.8875C30.0313 5.60625 32.3 3.7625 35 3.7625C35.2188 3.7625 36.4688 3.78125 37.625 4.3875C38.4373 4.79116 39.11 5.42902 39.5562 6.21875C39.9375 6.88125 40.0062 7.725 40.0062 8.0625C40.0062 8.4 39.8375 8.89375 39.5187 9.30625L39.525 9.3Z"
+            fill="url(#paint8_radial_10402_51337)"
+          />
+          <path
+            d="M39.525 9.3C39.4688 9.3625 39.3125 9.45625 39.3125 9.65C39.3125 9.8125 39.4187 9.975 39.6125 10.1062C40.5062 10.7312 42.2 10.6437 42.2062 10.6437C42.8725 10.6433 43.5265 10.4642 44.1 10.125C44.6774 9.78729 45.1565 9.30452 45.4897 8.72457C45.823 8.14463 45.9989 7.48764 46 6.81875C46.0187 5.41875 45.5 4.4875 45.2937 4.075C43.9625 1.4875 41.1062 -1.4243e-07 38 -9.16404e-08C35.8976 -0.000207945 33.8797 0.827186 32.3826 2.30323C30.8855 3.77928 30.0296 5.78532 30 7.8875C30.0313 5.60625 32.3 3.7625 35 3.7625C35.2188 3.7625 36.4688 3.78125 37.625 4.3875C38.4373 4.79116 39.11 5.42902 39.5562 6.21875C39.9375 6.88125 40.0062 7.725 40.0062 8.0625C40.0062 8.4 39.8375 8.89375 39.5187 9.30625L39.525 9.3Z"
+            fill="url(#paint9_radial_10402_51337)"
           />
         </g>
         <defs>
           <filter
-            id="filter0_dddd_7107_23656"
+            id="filter0_dddd_10402_51337"
             x="-4.9996"
             y="12.398"
-            width="34.9025"
+            width="34.9035"
             height="31.9416"
             filterUnits="userSpaceOnUse"
             color-interpolation-filters="sRGB"
@@ -113,7 +119,7 @@ export class ExtensionProviderIcon {
             <feGaussianBlur stdDeviation="1.26214" />
             <feComposite in2="hardAlpha" operator="out" />
             <feColorMatrix type="matrix" values="0 0 0 0 0.137255 0 0 0 0 0.968627 0 0 0 0 0.866667 0 0 0 0.1 0" />
-            <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_7107_23656" />
+            <feBlend mode="normal" in2="BackgroundImageFix" result="effect1_dropShadow_10402_51337" />
             <feColorMatrix
               in="SourceAlpha"
               type="matrix"
@@ -124,7 +130,7 @@ export class ExtensionProviderIcon {
             <feGaussianBlur stdDeviation="5.53398" />
             <feComposite in2="hardAlpha" operator="out" />
             <feColorMatrix type="matrix" values="0 0 0 0 0.137255 0 0 0 0 0.968627 0 0 0 0 0.866667 0 0 0 0.3 0" />
-            <feBlend mode="normal" in2="effect1_dropShadow_7107_23656" result="effect2_dropShadow_7107_23656" />
+            <feBlend mode="normal" in2="effect1_dropShadow_10402_51337" result="effect2_dropShadow_10402_51337" />
             <feColorMatrix
               in="SourceAlpha"
               type="matrix"
@@ -135,7 +141,7 @@ export class ExtensionProviderIcon {
             <feGaussianBlur stdDeviation="0.533981" />
             <feComposite in2="hardAlpha" operator="out" />
             <feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.35 0" />
-            <feBlend mode="normal" in2="effect2_dropShadow_7107_23656" result="effect3_dropShadow_7107_23656" />
+            <feBlend mode="normal" in2="effect2_dropShadow_10402_51337" result="effect3_dropShadow_10402_51337" />
             <feColorMatrix
               in="SourceAlpha"
               type="matrix"
@@ -146,11 +152,11 @@ export class ExtensionProviderIcon {
             <feGaussianBlur stdDeviation="0.0970874" />
             <feComposite in2="hardAlpha" operator="out" />
             <feColorMatrix type="matrix" values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.25 0" />
-            <feBlend mode="normal" in2="effect3_dropShadow_7107_23656" result="effect4_dropShadow_7107_23656" />
-            <feBlend mode="normal" in="SourceGraphic" in2="effect4_dropShadow_7107_23656" result="shape" />
+            <feBlend mode="normal" in2="effect3_dropShadow_10402_51337" result="effect4_dropShadow_10402_51337" />
+            <feBlend mode="normal" in="SourceGraphic" in2="effect4_dropShadow_10402_51337" result="shape" />
           </filter>
           <linearGradient
-            id="paint0_linear_7107_23656"
+            id="paint0_linear_10402_51337"
             x1="20.2306"
             y1="19.1846"
             x2="20.2306"
@@ -161,7 +167,7 @@ export class ExtensionProviderIcon {
             <stop offset="1" stop-color="#00FFE0" />
           </linearGradient>
           <linearGradient
-            id="paint1_linear_7107_23656"
+            id="paint1_linear_10402_51337"
             x1="6.49826"
             y1="30.7502"
             x2="42.1154"
@@ -172,7 +178,7 @@ export class ExtensionProviderIcon {
             <stop offset="1" stop-color="#4CFFE9" />
           </linearGradient>
           <linearGradient
-            id="paint2_linear_7107_23656"
+            id="paint2_linear_10402_51337"
             x1="20.2306"
             y1="25.3497"
             x2="20.2306"
@@ -183,7 +189,7 @@ export class ExtensionProviderIcon {
             <stop offset="1" stop-color="#00FFE0" />
           </linearGradient>
           <linearGradient
-            id="paint3_linear_7107_23656"
+            id="paint3_linear_10402_51337"
             x1="6.49826"
             y1="36.9152"
             x2="42.1154"
@@ -194,39 +200,79 @@ export class ExtensionProviderIcon {
             <stop offset="1" stop-color="#4CFFE9" />
           </linearGradient>
           <linearGradient
-            id="paint4_linear_7107_23656"
-            x1="31.0723"
-            y1="5.00001"
-            x2="44.9269"
-            y2="5.00001"
+            id="paint4_linear_10402_51337"
+            x1="33.666"
+            y1="11.0625"
+            x2="44.816"
+            y2="11.0625"
             gradientUnits="userSpaceOnUse"
           >
-            <stop stop-color="#D93025" />
-            <stop offset="1" stop-color="#EA4335" />
+            <stop stop-color="#0C59A4" />
+            <stop offset="1" stop-color="#114A8B" />
           </linearGradient>
+          <radialGradient
+            id="paint5_radial_10402_51337"
+            cx="0"
+            cy="0"
+            r="1"
+            gradientUnits="userSpaceOnUse"
+            gradientTransform="translate(39.8223 11.1466) scale(5.9625 5.66438)"
+          >
+            <stop offset="0.7" stop-opacity="0" />
+            <stop offset="0.9" stop-opacity="0.5" />
+            <stop offset="1" />
+          </radialGradient>
           <linearGradient
-            id="paint5_linear_7107_23656"
-            x1="36.9065"
-            y1="15.893"
-            x2="43.8339"
-            y2="3.89451"
+            id="paint6_linear_10402_51337"
+            x1="39.5417"
+            y1="6.2251"
+            x2="32.5855"
+            y2="13.8063"
             gradientUnits="userSpaceOnUse"
           >
-            <stop stop-color="#FCC934" />
-            <stop offset="1" stop-color="#FBBC04" />
+            <stop stop-color="#1B9DE2" />
+            <stop offset="0.2" stop-color="#1595DF" />
+            <stop offset="0.7" stop-color="#0680D7" />
+            <stop offset="1" stop-color="#0078D4" />
           </linearGradient>
-          <linearGradient
-            id="paint6_linear_7107_23656"
-            x1="38.8655"
-            y1="15.5005"
-            x2="31.9381"
-            y2="3.50195"
+          <radialGradient
+            id="paint7_radial_10402_51337"
+            cx="0"
+            cy="0"
+            r="1"
             gradientUnits="userSpaceOnUse"
+            gradientTransform="translate(34.4077 12.4337) rotate(-81.3844) scale(8.96163 7.24009)"
           >
-            <stop stop-color="#1E8E3E" />
-            <stop offset="1" stop-color="#34A853" />
-          </linearGradient>
-          <clipPath id="clip0_7107_23656">
+            <stop offset="0.8" stop-opacity="0" />
+            <stop offset="0.9" stop-opacity="0.5" />
+            <stop offset="1" />
+          </radialGradient>
+          <radialGradient
+            id="paint8_radial_10402_51337"
+            cx="0"
+            cy="0"
+            r="1"
+            gradientUnits="userSpaceOnUse"
+            gradientTransform="translate(31.6181 2.95725) rotate(92.2906) scale(12.6601 26.9634)"
+          >
+            <stop stop-color="#35C1F1" />
+            <stop offset="0.1" stop-color="#34C1ED" />
+            <stop offset="0.2" stop-color="#2FC2DF" />
+            <stop offset="0.3" stop-color="#2BC3D2" />
+            <stop offset="0.7" stop-color="#36C752" />
+          </radialGradient>
+          <radialGradient
+            id="paint9_radial_10402_51337"
+            cx="0"
+            cy="0"
+            r="1"
+            gradientUnits="userSpaceOnUse"
+            gradientTransform="translate(45.0038 4.83125) rotate(73.7398) scale(6.08125 4.94527)"
+          >
+            <stop stop-color="#66EB6E" />
+            <stop offset="1" stop-color="#66EB6E" stop-opacity="0" />
+          </radialGradient>
+          <clipPath id="clip0_10402_51337">
             <rect width="16" height="16" fill="white" transform="translate(30)" />
           </clipPath>
         </defs>