+++ src/components/common/provider-idle-screen/provider-idle-screen.tsx
@@ -3,7 +3,6 @@ import { Component, Event, Fragment, h, Prop } from '@stencil/core';
 import { getProviderButtonIcon } from 'components/functional/unlock-panel/helpers';
 import type { IProviderBase } from 'types/provider.types';
 import { ProviderTypeEnum } from 'types/provider.types';
-import { isFirefox } from 'utils/isFirefox';
 
 const getProviderIntroText = (providerType?: IProviderBase['type']) => {
   switch (providerType) {
@@ -46,7 +45,9 @@ export class ProviderIdleScreen {
     const extensionProviderIconWidth = extensionProviderIconBaseSize + (15 / 100) * extensionProviderIconBaseSize;
     const extensionProviderIconHeight = extensionProviderIconBaseSize + (10 / 100) * extensionProviderIconBaseSize;
 
-    const providerIntroIcon = getProviderButtonIcon(providerType);
+    const providerIntroIcon = isExtensionProvider
+      ? getProviderButtonIcon({ providerType, extensionProviderIconWidth, extensionProviderIconHeight })
+      : getProviderButtonIcon({ providerType });
     const providerIntroText = this.introText || getProviderIntroText(providerType);
 
     if (this.provider.type === ProviderTypeEnum.ledger) {
@@ -72,13 +73,7 @@ export class ProviderIdleScreen {
         />
 
         <div class="unlock-provider-intro">
-          {isExtensionProvider && !isFirefox() ? (
-            <div class="unlock-provider-intro-icon">
-              <mvx-extension-provider-icon width={extensionProviderIconWidth} height={extensionProviderIconHeight} />
-            </div>
-          ) : (
-            <div class="unlock-provider-intro-icon">{providerIntroIcon}</div>
-          )}
+          <div class="unlock-provider-intro-icon">{providerIntroIcon}</div>
 
           <div class="unlock-provider-intro-title">{this.introTitle}</div>
           {providerIntroText && <div class="unlock-provider-intro-text">{providerIntroText}</div>}