+++ src/constants/browser.enum.ts
@@ -1,0 +1,7 @@
+export enum BrowserEnum {
+  Chrome = 'Chrome',
+  Firefox = 'Firefox',
+  Edge = 'Edge',
+  Brave = 'Brave',
+  Arc = 'Arc',
+}